requirements.txt
chardet==5.2.0
colorama==0.4.6
et_xmlfile==2.0.0
importlib_metadata==8.6.1
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.0
reportlab==4.3.1
segno==1.6.6
six==1.17.0
tzdata==2025.2
zipp==3.21.0
Flask==2.3.3
Werkzeug==2.3.7


gunicorn==21.2.0
gevent==23.9.1
python-dotenv==1.0.0
Flask-Limiter==3.5.0
APScheduler==3.10.4
boto3==1.28.53
